Dockerized, added user/pass support
this is not working yet, but I needed to commit this pile of changes! the primary goal here is to be able to deploy with docker via docker-compose.yaml as part of home assistant setup.
I added user/pass support for MQTT, but I think it has also been added to the original repo since I created the fork, so I need to merge those changes
bunch of README changes to hopefully clarify a few things